Time: 7 ms (96.32%), Space: 6 MB (62.83%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
        long int rev=0;
-        int n=x;
+        long int n=x;
         if(x<0)
         {
             return false;
